added Section field migrated Checkbox field to Checkboxes - plural modifed Checkbox field to be single checkbox Replace Api class with Provider
