+++ src/Field/Section.php
@@ -1,23 +1,16 @@
 <?php
 namespace Shazzad\WpFormUi\Field;
 
-class Checkbox extends Field {
-
+class Section extends Field {
 	public function __construct( $data = [] ) {
-		$data['type'] = 'checkbox';
+		$data['type'] = 'section';
+
 		parent::__construct( $data );
 	}
 
 	public function get_html( $form ) {
 		$data = $this->sanitize_data( $this->data );
 		extract( $data );
-
-		if ( ! isset( $input_value ) ) {
-			$input_value = 'yes';
-		}
-		if ( ! isset( $input_label ) ) {
-			$input_label = $desc;
-		}
 
 		$html = $before;
 
@@ -42,17 +35,26 @@ class Checkbox extends Field {
 			$html .= sprintf( '<div class="%1$s %2$s"%3$s>', $this->form_pitc_class( 'wf-field-input-wrap', $id, $type ), $input_wrap_class, $input_wrap_attr );
 		}
 
-		$html .= $input_before;
+		// $html .= $input_before;
+		// $html .= sprintf(
+		// 	'<input class="%1$s %5$s" id="%2$s" name="%3$s" value="%4$s" type="%7$s"%6$s />',
+		// 	$this->form_pitc_class( 'wf-field', $id, $type ), $id, $name, $value, $input_class, $input_attr, $type
+		// );
+		// $html .= $input_after;
+		foreach ( $this->data['fields'] as $field ) {
+			$field = $form->create_field( $field );
+			if ( isset( $field->name ) && '' != $field->name && ! isset( $field->value ) ) {
+				$name = isset( $field->option_name ) ? $field->option_name : $field->name;
+				if ( array_key_exists( $name, $form->values ) ) {
+					$field->value = $form->values[ $name ];
+				} else {
+					$field->value = '';
+				}
+			}
 
-		$checked = ! empty( $value ) && $value === $input_value ? ' checked="checked"' : '';
-		$html .= sprintf(
-			'<label for="%1$s">
-				<input id="%1$s" name="%2$s" value="%3$s" type="checkbox"%4$s%6$s /> %5$s
-			</label>',
-			$id, $name, $input_value, $checked, $input_label, $input_attr
-		);
+			$html .= $field->get_html( $form );
+		}
 
-		$html .= $input_after;
 		if ( $input_wrap ) {
 			$html .= '</div>';
 		}
@@ -61,11 +63,7 @@ class Checkbox extends Field {
 
 		if ( isset( $desc_after ) ) {
 			if ( ! empty( $desc_after ) ) {
-				$html .= sprintf(
-					'<div class="%1$s">%2$s</div>',
-					$this->form_pitc_class( 'wf-field-desc-after-wrap', $id, $type ),
-					$desc_after
-				);
+				$html .= sprintf( '<div class="%1$s">%2$s</div>', $this->form_pitc_class( 'wf-field-desc-after-wrap', $id, $type ), $desc_after );
 			}
 		}
