+++ src/Form/Form.php
@@ -1,186 +1,149 @@
 <?php
 namespace Shazzad\WpFormUi\Form;
 
-abstract class Form implements \ArrayAccess {
-	public $data = array(
-		'settings' => array(),
-		'values'   => array(),
-		'fields'   => array(),
-		'rendered' => false
-	);
-	public $conditional_fields = [];
-
-	public function __construct( $data = array() ) {
-		if ( isset( $data['settings'] ) ) {
-			$this->set_settings( $data['settings'] );
-		}
-		if ( isset( $data['values'] ) ) {
-			$this->set_values( $data['values'] );
-		}
-		if ( isset( $data['fields'] ) ) {
-			$this->add_fields( $data['fields'] );
-		}
-	}
-	public function set_settings( $settings = array() ) {
-		$this->settings = $settings;
-	}
-	public function set_values( $values = array() ) {
-		$this->values = $values;
-	}
-	public function add_fields( $fields ) {
-		foreach ( $fields as $field ) {
-			$this->add_field( $field );
-		}
+class Form extends Base {
+	function __construct( $data = array() ) {
+		parent::__construct( $data );
 	}
 
-	public function set_setting( $key, $val = null ) {
-		$this->settings[ $key ] = $val;
-	}
-	public function set_value( $key, $val = null ) {
-		$this->values[ $key ] = $val;
-	}
-	public function add_field( $data ) {
-		$this->data['fields'][] = $this->create_field( $data );
-	}
-	public function create_field( $data ) {
-		if ( ! isset( $data['type'] ) ) {
-			$data['type'] = 'html';
+	public function toHtml() {
+		if ( ! is_array( $this->fields ) ) {
+			$this->fields = array();
 		}
-		if ( ! isset( $data['priority'] ) ) {
-			$data['priority'] = 10;
+		if ( ! is_array( $this->values ) ) {
+			$this->values = array();
 		}
-		$class_name = '\\Shazzad\WpFormUi\\Field\\' . str_replace( ' ', '_', ucwords( str_replace( '_', ' ', $data['type'] ) ) );
-		if ( ! class_exists( $class_name ) ) {
-			$class_name = '\\Shazzad\WpFormUi\\Field\\Html';
+		if ( ! is_array( $this->settings ) ) {
+			$this->settings = array();
 		}
 
-		return new $class_name( $data, $this );
-	}
-	public function render() {
-		if ( ! $this->rendered ) {
-			$this->handle_fields_conditional_logics();
-			uasort( $this->data['fields'], array( $this, 'order_by_priority' ) );
-			$this->rendered = true;
+		if ( empty( $this->settings['method'] ) ) {
+			$this->settings['method'] = 'POST';
+		}
+		if ( empty( $this->settings['class'] ) ) {
+			$this->settings['class'] = 'wf wf-basic';
+		} else {
+			$this->settings['class'] = 'wf ' . $this->settings['class'];
+		}
+		if ( ! empty( $this->settings['ajax'] ) ) {
+			$this->settings['class'] .= ' wf-ajax';
+		}
+		if ( ! empty( $this->settings['ajax'] ) && empty( $this->settings['loading_text'] ) ) {
+			$this->settings['loading_text'] = 'Updating';
+		}
+		if ( empty( $this->settings['button_text'] ) ) {
+			$this->settings['button_text'] = 'Update';
+		}
+		// @since v:2.0
+		if ( empty( $this->settings['attr'] ) ) {
+			$this->settings['attr'] = '';
+		}
+		if ( ! empty( $this->settings['loading_text'] ) ) {
+			$this->settings['attr'] .= ' data-loading_text="' . esc_attr( $this->settings['loading_text'] ) . '"';
+		}
+		if ( ! empty( $this->settings['success_text'] ) ) {
+			$this->settings['attr'] .= ' data-success_text="' . esc_attr( $this->settings['success_text'] ) . '"';
+		}
+		if ( ! empty( $this->settings['context'] ) ) {
+			$this->settings['class'] .= ' wf-field-context-' . $this->settings['context'];
 		}
 
-		echo $this->toHtml();
-		echo $this->formJs();
-	}
+		if ( empty( $this->settings['action'] ) ) {
+			$this->settings['action'] = ( is_ssl() ? 'https://' : 'http://' ) . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
+		}
 
-	public function formJs() {
-		?>
-		<script type="application/javascript">
-			if (!window['wf_form_conditional_logic']) {
-				window['wf_form_conditional_logic'] = {};
-			}
-			window['wf_form_conditional_logic']["<?php echo $this->settings['id']; ?>"] = <?php echo json_encode( $this->conditional_fields ); ?>;
-		</script>
-		<?php
-	}
-
-	public function handle_fields_conditional_logics() {
-		$this->conditional_fields = [];
-		$field_conditions         = [];
-		foreach ( $this->data['fields'] as $field ) {
-			if ( isset( $field['conditional_logics'] ) ) {
-				$conditional_logics                        = $field['conditional_logics'];
-				$this->conditional_fields[ $field['key'] ] = $conditional_logics;
-
-				foreach ( $conditional_logics['rules'] as $rule ) {
-					$field_key = $rule['key'];
-					if ( ! isset( $field_conditions[ $field_key ] ) ) {
-						$field_conditions[ $field_key ] = [ $field['key'] ];
-					} else {
-						$field_conditions[ $field_key ][] = $field['key'];
-					}
+		// query variables
+		if ( ! empty( $this->settings['qv'] ) ) {
+			$query_vars = array();
+			foreach ( $this->settings['qv'] as $q ) {
+				if ( isset( $_GET[ $q ] ) && $_GET[ $q ] != '' ) {
+					$query_vars[ $q ] = trim( $_GET[ $q ] );
 				}
 			}
-		}
 
-		foreach ( $field_conditions as $field_key => $fields ) {
-			$index = $this->get_field_index( $field_key );
-			if ( ! isset( $this->data['fields'][ $index ]['input_attrs'] ) ) {
-				$this->data['fields'][ $index ]['input_attrs'] = [ 
-					'onchange' => 'wf_apply_rules("' . $this->settings['id'] . '",' . json_encode( $fields ) . ');',
-					//'onkeyup' => 'clearTimeout(__wf_timeout_handle); __wf_timeout_handle = setTimeout("wf_apply_rules("'.$this->settings['id'].'",'.json_encode($fields).')", 300);'
-				];
-			} else {
-				$this->data['fields'][ $index ]['input_attrs']['onchange'] = 'wf_apply_rules("' . $this->settings['id'] . '",' . json_encode( $fields ) . ');';
-			}
-
-			if ( isset( $this->data['fields'][ $index ]['class'] ) ) {
-				$this->data['fields'][ $index ]['class'] .= ' has-logic';
-			} else {
-				$this->data['fields'][ $index ]['class'] = 'has-logic';
+			if ( ! empty( $query_vars ) ) {
+				$this->settings['action'] = add_query_arg( $query_vars, $this->settings['action'] );
 			}
 		}
-		#print_r($this->data['fields']);
-		#die();
-	}
-	public function get_field_index( $key ) {
-		foreach ( $this->data['fields'] as $i => $field ) {
-			if ( $field['key'] == $key ) {
-				return $i;
+
+		$html = '';
+
+		// form opening tag
+		if ( ! isset( $this->settings['no_form'] ) ) {
+			$html .= '<form';
+			$attr_keys = array( 'class', 'id', 'name', 'title', 'enctype', 'method', 'action' );
+			foreach ( $this->settings as $name => $attr ) {
+				if ( ! empty( $name ) && in_array( $name, $attr_keys ) ) {
+					$html .= ' ' . $name . '="' . esc_attr( $attr ) . '"';
+				}
 			}
-		}
-		return false;
-	}
-
-	public function order_by_priority( $a, $b ) {
-		if ( ! isset( $a['priority'] ) || ! isset( $b['priority'] ) )
-			return -1;
-		if ( $a['priority'] == $b['priority'] )
-			return 0;
-		return ( $a['priority'] < $b['priority'] ) ? -1 : 1;
-	}
-
-	// usability
-	public function toHtml() {
-		return '';
-	}
-	public function toArray() {
-		return $this->data;
-	}
-	public function toJson() {
-		return json_encode( $this->data );
-	}
-	// no magic
-	public function __sleep() {
-		return array_keys( $this->data );
-	}
-	public function __wakeup() {
-	}
-	public function __toString() {
-		#echo 'field '. $this->data['type'];
-		return $this->toHtml();
-	}
-	public function &__get( $key ) {
-		return $this->data[ $key ];
-	}
-	public function __set( $key, $value ) {
-		$this->data[ $key ] = $value;
-	}
-	public function __isset( $key ) {
-		return isset( $this->data[ $key ] );
-	}
-	public function __unset( $key ) {
-		unset( $this->data[ $key ] );
-	}
-	// array access
-	public function offsetGet( $offset ) {
-		return isset( $this->data[ $offset ] ) ? $this->data[ $offset ] : null;
-	}
-	public function offsetSet( $offset, $value ) {
-		if ( is_null( $offset ) ) {
-			$this->data[] = $value;
+			if ( ! empty( $this->settings['attr'] ) ) {
+				$html .= $this->settings['attr'];
+			}
+			$html .= '>';
 		} else {
-			$this->data[ $offset ] = $value;
+			$html .= '<div';
+			$attr_keys = array( 'class', 'id' );
+			foreach ( $this->settings as $name => $attr ) {
+				if ( ! empty( $name ) && in_array( $name, $attr_keys ) ) {
+					$html .= ' ' . $name . '="' . esc_attr( $attr ) . '"';
+				}
+			}
+			if ( ! empty( $this->settings['attr'] ) ) {
+				$html .= $this->settings['attr'];
+			}
+			$html .= '>';
 		}
-	}
-	public function offsetExists( $offset ) {
-		return isset( $this->data[ $offset ] );
-	}
-	public function offsetUnset( $offset ) {
-		unset( $this->data[ $offset ] );
+
+		if ( ! empty( $this->settings['after_tag'] ) ) {
+			$html .= $this->settings['after_tag'];
+		}
+
+		if ( ! empty( $this->settings['title'] ) ) {
+			$html .= '<div class="wf-form-title">' . $this->settings['title'] . '</div>';
+		}
+
+		// Submit button on top of the form
+		if ( isset( $this->settings['submit_top'] ) && $this->settings['submit_top'] === true ) {
+			$html .= "<div class='wf-field-wrap wf-field-wrap-type-submit wf-field-wrap-type-submit-top'>
+				<input type='submit' value='" . $this->settings['button_text'] . "' class='button button-primary form_button button_top'>
+			</div>";
+		}
+
+		foreach ( $this->fields as &$field ) {
+			if ( isset( $field->name ) && '' != $field->name && ! isset( $field->value ) ) {
+				$name = isset( $field->option_name ) ? $field->option_name : $field->name;
+				if ( array_key_exists( $name, $this->values ) ) {
+					$field->value = $this->values[ $name ];
+				} else {
+					$field->value = '';
+				}
+			}
+
+			if ( empty( $field->input_class ) && ! empty( $field->name ) && 'action' == $field->name ) {
+				$field->input_class = 'form_action';
+			}
+
+			$html .= $field->get_html( $this );
+		}
+
+		// Submit button at the end of the form
+		if ( ! isset( $this->settings['submit_bottom'] ) || $this->settings['submit_bottom'] !== false ) {
+			$html .= "<div class='wf-field-wrap wf-field-wrap-type-submit wf-field-wrap-type-submit-bottom'>
+				<input type='submit' value='" . $this->settings['button_text'] . "' class='form_button button_bottom'>
+			</div>";
+		}
+
+		if ( ! empty( $this->settings['form_closing'] ) ) {
+			$html .= $this->settings['form_closing'];
+		}
+
+		if ( ! isset( $this->settings['no_form'] ) ) {
+			$html .= '</form>';
+		} else {
+			$html .= '</div>';
+		}
+
+		return $html;
 	}
 }
